+++ CHANGES.rst
@@ -1,6 +1,25 @@
 Changelog
 =========
 
+1.0b1 (2012-01-27)
+------------------
+
+- Added mime type icon for file.
+  [loechel]
+
+- Lead image behavior added.
+  [timo]
+
+- Make NewsItem use the lead image behavior.
+  [timo]
+
+- SearchableText indexes added.
+  [reinhardt]
+
+- Robot framework test added.
+  [Gomez]
+
+
 1.0a2 (unreleased)
 ------------------
 
@@ -13,8 +32,8 @@ Changelog
 - Migration view (@@fix_base_classes) added to migrate content objects that
   were created with version 1.0a1.
   [timo]
-  
-- Mime Type Icon added for File View [loechel] 
+
+- Mime Type Icon added for File View [loechel]
 
 
 1.0a1 (unreleased)